move db to infra directory
